dashboard/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
boto3>=1.34.0
scikit-learn>=1.4.0
textblob>=0.18.0
vaderSentiment>=3.3.2
statsmodels>=0.14.0
pytest-json-report>=1.5.0
